+++ lib/form/index.js
@@ -15,8 +15,9 @@ module.exports = {
     mail: require('./admin/mail'),
     aws: require('./admin/aws'),
     plugin: require('./admin/plugin'),
-    securityGeneral: require('./admin/securitygeneral'),
-    securityGoogle: require('./admin/securitygoogle'),
+    securityGeneral: require('./admin/securityGeneral'),
+    securityGoogle: require('./admin/securityGoogle'),
+    securityMechanism: require('./admin/securityMechanism'),
     markdown: require('./admin/markdown'),
     customcss: require('./admin/customcss'),
     customscript: require('./admin/customscript'),
